net.talaatharb.projecteuler100.Core
