--- csharp-project-euler-100/csharp-project-euler-100.slnx
@@ -1,3 +1,0 @@
-<Solution>
-  <Project Path="csharp-project-euler-100.csproj" />
-</Solution>
