+++ csharp-project-euler-100/net.talaatharb.projecteuler100.Core/P006.cs
@@ -1,6 +1,6 @@
 
 
-namespace net.talaatharb.projecteuler100
+namespace net.talaatharb.projecteuler100.Core
 {
 
     /**
