+++ csharp-project-euler-100/net.talaatharb.projecteuler100.Core/P012.cs
@@ -1,13 +1,13 @@
 
 
-namespace net.talaatharb.projecteuler100
+namespace net.talaatharb.projecteuler100.Core
 {
 
     /**
     * @author mharb
     *
     */
-    public class P011 : Solution
+    public class P012 : Solution
     {
 
         /**
@@ -17,12 +17,12 @@ namespace net.talaatharb.projecteuler100
          */
         public static void Main(string[] args)
         {
-            new P011().Run();
+            new P012().Run();
         }
 
         public override long Solve()
         {
-            problemNumber = 11;
+            problemNumber = 12;
             return problemNumber;
         }
     }
