+++ csharp-project-euler-100/net.talaatharb.projecteuler100.Core/P014.cs
@@ -1,13 +1,13 @@
 
 
-namespace net.talaatharb.projecteuler100
+namespace net.talaatharb.projecteuler100.Core
 {
 
     /**
     * @author mharb
     *
     */
-    public class P012 : Solution
+    public class P014 : Solution
     {
 
         /**
@@ -17,12 +17,12 @@ namespace net.talaatharb.projecteuler100
          */
         public static void Main(string[] args)
         {
-            new P012().Run();
+            new P014().Run();
         }
 
         public override long Solve()
         {
-            problemNumber = 12;
+            problemNumber = 14;
             return problemNumber;
         }
     }
