+++ csharp-project-euler-100/net.talaatharb.projecteuler100.Core/Solution.cs
@@ -1,6 +1,6 @@
 using System.Diagnostics;
 
-namespace net.talaatharb.projecteuler100
+namespace net.talaatharb.projecteuler100.Core
 {
 
     /**
