+++ csharp-project-euler-100/net.talaatharb.projecteuler100.Core/Solvable.cs
@@ -1,5 +1,5 @@
 
-namespace net.talaatharb.projecteuler100
+namespace net.talaatharb.projecteuler100.Core
 {
 
     /**
