+++ csharp-project-euler-100/net.talaatharb.projecteuler100.slnx
@@ -1,0 +1,4 @@
+<Solution>
+  <Project Path="csharp-project-euler-100.csproj" />
+  <Project Path="net.talaatharb.projecteuler100.Core/net.talaatharb.projecteuler100.Core.csproj" Id="d83d218d-f492-41b0-89d0-2c91bcd1d4b8" />
+</Solution>
