+++ csharp-project-euler-100/src/AllSolutions.cs
@@ -1,4 +1,4 @@
-﻿namespace net.talaatharb.projecteuler100
+﻿namespace net.talaatharb.projecteuler100.Core
 {
     public class AllSolutions
     {
